+++ clients/derivatives_trading_portfolio_margin/src/binance_sdk_derivatives_trading_portfolio_margin/rest_api/rest_api.py
@@ -173,6 +173,7 @@ from .models import NewUmAlgoOrderPositionSideEnum
 from .models import NewUmAlgoOrderTimeInForceEnum
 from .models import NewUmAlgoOrderWorkingTypeEnum
 from .models import NewUmAlgoOrderPriceMatchEnum
+from .models import NewUmAlgoOrderClosePositionEnum
 from .models import NewUmAlgoOrderPriceProtectEnum
 from .models import NewUmAlgoOrderReduceOnlyEnum
 from .models import NewUmAlgoOrderNewOrderRespTypeEnum
@@ -2821,13 +2822,14 @@ class DerivativesTradingPortfolioMarginRestAPI:
         symbol: Union[str, None],
         side: Union[NewUmAlgoOrderSideEnum, None],
         type: Union[NewUmAlgoOrderTypeEnum, None],
-        quantity: Union[float, None],
         position_side: Optional[NewUmAlgoOrderPositionSideEnum] = None,
         time_in_force: Optional[NewUmAlgoOrderTimeInForceEnum] = None,
+        quantity: Optional[float] = None,
         price: Optional[float] = None,
         trigger_price: Optional[float] = None,
         working_type: Optional[NewUmAlgoOrderWorkingTypeEnum] = None,
         price_match: Optional[NewUmAlgoOrderPriceMatchEnum] = None,
+        close_position: Optional[NewUmAlgoOrderClosePositionEnum] = None,
         price_protect: Optional[NewUmAlgoOrderPriceProtectEnum] = None,
         reduce_only: Optional[NewUmAlgoOrderReduceOnlyEnum] = None,
         activate_price: Optional[float] = None,
@@ -2856,6 +2858,7 @@ class DerivativesTradingPortfolioMarginRestAPI:
         - `STOP`, `STOP_MARKET`: BUY: latest price >= `triggerPrice`; SELL: latest price <= `triggerPrice`.
         - `TAKE_PROFIT`, `TAKE_PROFIT_MARKET`: BUY: latest price <= `triggerPrice`; SELL: latest price >= `triggerPrice`.
         - `TRAILING_STOP_MARKET`: BUY: lowest price after order placed <= `activatePrice`, and latest price >= lowest price * (1 + `callbackRate`); SELL: highest price after order placed >= `activatePrice`, and latest price <= highest price * (1 - `callbackRate`).
+        - `STOP_MARKET`, `TAKE_PROFIT_MARKET` with `closePosition`=`true`: follow the same rules for conditional orders; if triggered, close all current long position (if `SELL`) or current short position (if `BUY`); cannot be used with `quantity` parameter; cannot be used with `reduceOnly` parameter; in Hedge Mode, cannot be used with `BUY` orders in `LONG` position side, and cannot be used with `SELL` orders in `SHORT` position side.
         - `selfTradePreventionMode` is only effective when `timeInForce` set to `IOC` or `GTC` or `GTD`.
 
                 Args:
@@ -2863,15 +2866,16 @@ class DerivativesTradingPortfolioMarginRestAPI:
                     symbol (Union[str, None]):
                     side (Union[NewUmAlgoOrderSideEnum, None]):
                     type (Union[NewUmAlgoOrderTypeEnum, None]): Conditional order type
-                    quantity (Union[float, None]): Order quantity
                     position_side (Optional[NewUmAlgoOrderPositionSideEnum] = None): Default `BOTH` for One-way Mode; `LONG` or `SHORT` for Hedge Mode
                     time_in_force (Optional[NewUmAlgoOrderTimeInForceEnum] = None):
+                    quantity (Optional[float] = None): Order quantity. Cannot be sent with `closePosition`=`true`(Close-All)
                     price (Optional[float] = None): Order price
                     trigger_price (Optional[float] = None): Trigger price
                     working_type (Optional[NewUmAlgoOrderWorkingTypeEnum] = None): Trigger price type. Default `CONTRACT_PRICE`
                     price_match (Optional[NewUmAlgoOrderPriceMatchEnum] = None): Can't be passed together with `price`
+                    close_position (Optional[NewUmAlgoOrderClosePositionEnum] = None): Close-All, used with `STOP_MARKET` or `TAKE_PROFIT_MARKET`.
                     price_protect (Optional[NewUmAlgoOrderPriceProtectEnum] = None): Price protection. Default `false`
-                    reduce_only (Optional[NewUmAlgoOrderReduceOnlyEnum] = None): Cannot be sent in Hedge Mode
+                    reduce_only (Optional[NewUmAlgoOrderReduceOnlyEnum] = None): Cannot be sent in Hedge Mode; cannot be sent with `closePosition`=`true`
                     activate_price (Optional[float] = None): Used with `TRAILING_STOP_MARKET`, default as latest price
                     callback_rate (Optional[float] = None): Used with `TRAILING_STOP_MARKET`, min 0.1, max 10 (1 = 1%)
                     client_algo_id (Optional[str] = None): Unique id among open orders. Auto-generated if not sent
@@ -2893,13 +2897,14 @@ class DerivativesTradingPortfolioMarginRestAPI:
             symbol,
             side,
             type,
-            quantity,
             position_side,
             time_in_force,
+            quantity,
             price,
             trigger_price,
             working_type,
             price_match,
+            close_position,
             price_protect,
             reduce_only,
             activate_price,